jni: added getUtxos method

+++ android/com/mw/beam/beamwallet/core/SystemState.java
@@ -3,5 +3,5 @@ package com.mw.beam.beamwallet.core;
 public class SystemState
 {
 	public byte[] hash;
-	public int height;
+	public long height;
 }
